+++ src/Yaat.Client/ViewModels/MainViewModel.cs
@@ -1380,6 +1380,7 @@ public partial class MainViewModel : ObservableObject
         _connection.AircraftUpdated += OnAircraftUpdated;
         _connection.AircraftDeleted += OnAircraftDeleted;
         _connection.AircraftSpawned += OnAircraftSpawned;
+        _connection.ScenarioRestarted += OnScenarioRestarted;
 
         _cfrExpiryTimer = new DispatcherTimer { Interval = TimeSpan.FromSeconds(1) };
         _cfrExpiryTimer.Tick += (_, _) => SweepCfrExpiry();